Mainscreen added
Mainscreen template added and works when correct credentials are added. This can serve as a baseplate to the rest of the application

I put it within the login package for now as I couldn't get it working from it's own package. however this can be changed later if we can figure it out. Otherwise we could alter the Login package name to Main or something for consistency and just stick everything in there.